bug fix in runVisStim

+++ runVisStim.m
@@ -27,7 +27,7 @@ function runVisStim()
     prompt = 'Duration to play visual stimulus (s): ';
     dur = input(prompt, 's');
     % convert duration to datetime format
-    durSec = seconds(dur);
+    durSec = seconds(double(dur));
     
     % prompt user for visual stimulus parameters
     visstimParams = visstimUserPrompts(settings);
